requirements.txt
Flask
Flask-Cors
scipy
numpy
opencv-python-headless
pillow==9.0.1
paddlepaddle==2.4.0
gunicorn==20.0.4
onnxruntime==1.12.0
pyclipper
shapely==2.0.7
urllib3==1.26.6
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84
